change input parsing & execution logic, todo tests
