+++ src/compute/mod.rs
@@ -5,4 +5,4 @@ mod nonce;
 pub use nonce::mine_nonce;
 
 mod workflows;
-pub use workflows::WorkflowsExt;
+pub use workflows::*;
